Move websocket to core module
