+++ tests/core/test_websocket.py
@@ -1,8 +1,9 @@
 # coding=utf-8
 
-import autosubliminal
-from autosubliminal.websocket import WebSocketHandler
 from ws4py.messaging import TextMessage, BinaryMessage
+
+import autosubliminal
+from autosubliminal.core.websocket import WebSocketHandler
 
 
 class MyProcess(object):
@@ -14,14 +15,14 @@ autosubliminal.SCHEDULERS = {'MyProcess': MyProcess()}
 
 
 def test_receive_text_message(mocker):
-    mocker.patch('autosubliminal.websocket.WebSocketHandler.handle_message', return_value=True)
+    mocker.patch('autosubliminal.core.websocket.WebSocketHandler.handle_message', return_value=True)
     handler = WebSocketHandler(None)
     handler.received_message(TextMessage('{"key1": "value1"}'))
     handler.handle_message.assert_called_once_with({'key1': 'value1'})
 
 
 def test_receive_invalid_message(mocker):
-    mocker.patch('autosubliminal.websocket.WebSocketHandler.handle_message')
+    mocker.patch('autosubliminal.core.websocket.WebSocketHandler.handle_message')
     handler = WebSocketHandler(None)
     handler.received_message(BinaryMessage(b''))
     handler.handle_message.assert_not_called()
